Move Troops and add Type
