+++ WarOfEmpires.Domain.Tests/Players/TroopsTests.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
+using WarOfEmpires.Domain.Attacks;
 using WarOfEmpires.Domain.Players;
 
 namespace WarOfEmpires.Domain.Tests.Players {
