+++ WarOfEmpires.Domain/WarOfEmpires.Domain.csproj
@@ -77,7 +77,7 @@
     <Compile Include="Players\ProductionInfo.cs" />
     <Compile Include="Empires\RecruitTaskTriggeredEvent.cs" />
     <Compile Include="Players\TroopInfo.cs" />
-    <Compile Include="Players\Troops.cs" />
+    <Compile Include="Attacks\Troops.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Events\ScheduledTask.cs" />
     <Compile Include="Security\Password.cs" />
